Revert "Replace openlayers control with simple cick handler function"
This reverts commit d7ace7c83461f91cc0e1d18a9752e602ffcf5670.

The reverted commit caused a following regression:
When a bus stop (asset) was first selected and then user clicked mouse button down and dragged the map, the bus stop was unselected when mouse button was released.

Earlier this did not happen cause the OpenLayers control based click handler was given a pixelTolerance option with value 0. The pixelTolerance option specifies a maximum number of pixels between mouseup and mousedown for an event to be considered a click, and hence dragging won't cause a click.