Fix speed limit selection by mml id from url
The layer change and moving the map may both result in backend bounding box fetch.

If the bounding box fetch resulting from the layer change finishes last,
then the collection can still be in a state where the speed limit is no more
in the collection.

Make the ordering explicit and wait until the layer change is finished and move the map only after that.