Use Bacon to enforce processing only the latest linear assets response.
The code is gnarly as hell because we're doing something fairly silly:
we start with a promise API (getLinearAssets), turn it into a Bacon
stream (`bus` and `responses`), and turn events in that stream into new
promises!

All calls of getLinearAssets need to share the same bus/stream where the
events go, because otherwise we wouldn't guarantee any relationship between
result values.

Please kill it with fire.