DROTH-252 Show pop-up if mass transit stop is moved more than 50 meters
Created a generic pop-up that can be customised to be used in many cases.

Edited the index.html template in order to include the new popup (grunt will generate the correct index.html).

Edited the MassTransitStopLayer in order to accommodate the inclusion of said popup.

Several assumptions were made:

The distance is calculated between the starting point of the bus stop and the current snappable point.
Said distance is calculated using the method getSquaredDistanceBetweenPoints from geometryCalculator and it is assumed that the result is given out in meters, if not there is a simple multiplier (distanceMult) that will allow the conversion between units.
The distance trigger for the pop up is 50 meters (via the movementLimit parameter).
If we give permission to move a bus stop then we can continue to move said bus stop until we de-select it, on further selection it will ask for permission again.