Historylink processing.
Work in progress

Goes through historylinks and compares them to current links.
if no matching historylink is found to current link  it is kept
if historylink has matching current link  with in tolerance it is kept
removed when historylink is not with in tolerance.

TODO:
Check tolerance values
Test to check it works as intended
cleanup