DRVVH-377 (335) UI: Display the complementary geometry on top (Z-indexing) in visualization rules
Changed the RoadLayer in order to accommodate for the newly added Z-Level based on the roadlink subtype.
Changed the draw dashed and borders in order to accommodate for the Z-Level in the LinkPropertyLayer.
Changed the LinkPropertyLayerStyles in order to find the zLevel of a roadLink and apply it.
Changed ViiteApi in order to send the subType on the return object.