DROTH-2029 unable to save floating stops on terminated roads
In the case of bus stop 307276:
It was located on link 4385749. There is changes made to this link, so it becomes one-directional and another road is added next to it. In Tierekisteri it is terminated. But in this case, it is keeping the old link Id, but the the road now has only one direction. In the code we assume that if the old link Id still exists, the road is not terminated. This will be changed to always check for both the current link and the historic one, because the Terminated strategy uses the historic one.