DROTH-4468 Assets on expired road links -worklist improvements.
New feature where assets on expired road links can be viewed on a pop up map.
Assets on the worklist are also expired when removed from the worklist.
Asset coordinates can be copied to the clipboard from the pop up map view.
Additional info of the asset can be copied to clipboard from the worklist.